Rename the main directory with AR algorithms code files
